re-organize query utils slightly
