+++ src/data/queries/event.ts
@@ -9,7 +9,10 @@ import { NodeEvent } from '@/types/drupal/node'
 import { Event } from '@/types/formatted/event'
 import { GetServerSidePropsContext } from 'next'
 import { MediaImage } from '@/types/formatted/media'
-import { entityBaseFields, fetchSingleEntityOrPreview } from '@/lib/utils/query'
+import {
+  entityBaseFields,
+  fetchSingleEntityOrPreview,
+} from '@/lib/drupal/queryResources'
 
 export const params: QueryParams<null> = () => {
   return queries
